Add a draw_hist routine to Hist, using matplotlib
This just uses the matplotlib histogram function for now, and all oddities that that entails. No Axis names or labels are included at this stage. This function is the only matplotlib reference so far, so as long as you don't try to call it matplotlib isn't strictly required.

It also defaults to using a step-type histogram, as I don't know anyone in my field that uses bar-type histograms. This may be a little sociopathic, so may be removed in the future.